requirements.txt
# Core data science and ML libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# Data processing and storage
pyarrow>=10.0.0

# API and web serving
fastapi>=0.100.0
uvicorn>=0.20.0

# Configuration and utilities
pyyaml>=6.0
pydantic>=1.10.0

# Testing
pytest>=7.0.0
httpx>=0.24.0  # For FastAPI test client

# Optional ML libraries (warn if unavailable)
# lightgbm>=3.3.0  # May not be available on Apple Silicon
# xgboost>=1.6.0   # May not be available on Apple Silicon

# Jupyter for exploration
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Development and deployment
docker>=6.0.0
pytest-asyncio>=0.21.0
